requirements.txt
python-dotenv
langchain
langchain-core
langchain-openai
fastapi
uvicorn
pydantic
python-multipart